==> spatial_HYP/code/10-Spatial Registration/03-Output spatreg supp tab sheets.Rmd <==
---
title: "03-Output spatreg supp tab sheets"
author: "Bernard Mulvey"
date: "2024-08-04"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(fig.height = 10,fig.width = 7,include = FALSE)
knitr::opts_chunk$set(fig.width=7,fig.height=10)
#### sets tab autocompletion for directories to begin from the directory containing the .Rproj session, instead of the script's directory if different
knitr::opts_knit$set(root.dir = here::here())

library(data.table)
library(Biostrings)
require(colorout)
ColorOut()
options("styler.addins_style_transformer" = "biocthis::bioc_style()")

## change write.table to output a TSV without quotes or rownames by default
library(default)
default(write.table) <- list(row.names=F,col.names=T,sep='\t',quote=F)
```

### save sccoco result sheets for k15, 20, 31, 15collapsed
```{r}
scc <- readRDS("processed-data/10-Spatial Registration/01b-scCoco query of top 150 markers per k15-20-31-collapsedK15 domain in ABA ms hyp CCF areas.RDS")

i <- 1
for (i in c(1:length(scc))){
   outtab <- as.data.table(scc[[i]][[1]])
   outname <- names(scc)[i]
   
   # tidy up colnames
   fixpat <- paste0(names(scc)[i],"_")
   setnames(outtab,gsub(names(outtab),pattern=fixpat,replacement=""))
   
   fwrite(outtab,paste0("local_supptab_assembly/sccoco-ABA_",outname,".txt"),sep='\t',quote=F)
}

rm(scc)

```


### writing out spatial reg results for yao23 ABA subclasses (we only want collapsed ARH and VMH subclasses for comparing to Visium ARC/VMH)
```{r}
spatregs <- readRDS("processed-data/10-Spatial Registration/02g-sphyp-k15-20-31-15clps_ABAms-47subclass-47withVMHARCclps-VMHARCsupertypes.RDS")

spatregs <- spatregs[grep(names(spatregs),pattern="subclass_collapsed")]

spatregs2 <- lapply(spatregs,FUN=function(x,n){
   ret <- mapply(d=x,n=names(x),SIMPLIFY=FALSE,FUN=function(d,n){
      z <- as.data.table(d)
      z[,ntop:=as.numeric(gsub(n,pattern="^.*_top_(.*)_sphypdomainmks$",replacement="\\1"))]
      z[,comparison:=n]
      return(z)
   })
   ret2 <- rbindlist(ret)
   return(ret2)
})

ret3 <- rbindlist(spatregs2)
ret3[,comparison:=gsub(comparison,pattern="ABA23_HYP_subclass_collapsed_to_(.*)_top.*$",replacement="\\1")]
unique(ret3$comparison)
ret3[,comparison:=gsub(comparison,pattern="sphyp",replacement="")]
ret3[,comparison:=gsub(comparison,pattern="_collapsed",replacement=" (collapsed VMH/ARC)")]
setnames(ret3,"comparison","BayesSpace Clustering")
setnames(ret3,"sphyp_domain","Visium Cluster")



## name ARC and VMH clusters for the nnsvg10 harmony lambda=NA bayesspaces
ret3[`BayesSpace Clustering`=="k15"&`Visium Cluster`=="X4",`Visium Cluster`:="ARC.1"]
ret3[`BayesSpace Clustering`=="k15"&`Visium Cluster`=="X6",`Visium Cluster`:="ARC.2"]
ret3[`BayesSpace Clustering`=="k15"&`Visium Cluster`=="X7",`Visium Cluster`:="VMH.1"]
ret3[`BayesSpace Clustering`=="k15"&`Visium Cluster`=="X12",`Visium Cluster`:="VMH.2"]

ret3[`BayesSpace Clustering`=="k20"&`Visium Cluster`=="X1",`Visium Cluster`:="ARC20.1"]
ret3[`BayesSpace Clustering`=="k20"&`Visium Cluster`=="X4",`Visium Cluster`:="ARC20.2"]
ret3[`BayesSpace Clustering`=="k20"&`Visium Cluster`=="X2",`Visium Cluster`:="VMH20.1"]
ret3[`BayesSpace Clustering`=="k20"&`Visium Cluster`=="X3",`Visium Cluster`:="VMH20.2"]

ret3[`BayesSpace Clustering`=="k31"&`Visium Cluster`=="X6",`Visium Cluster`:="VMH31.1"]
ret3[`BayesSpace Clustering`=="k31"&`Visium Cluster`=="X21",`Visium Cluster`:="VMH31.2"]
ret3[`BayesSpace Clustering`=="k31"&`Visium Cluster`=="X25",`Visium Cluster`:="VMH31.3"]
ret3[`BayesSpace Clustering`=="k31"&`Visium Cluster`=="X26",`Visium Cluster`:="VMH31.4"]
ret3[`BayesSpace Clustering`=="k31"&`Visium Cluster`=="X31",`Visium Cluster`:="VMH31.5"]
ret3[`BayesSpace Clustering`=="k31"&`Visium Cluster`=="X2",`Visium Cluster`:="ARC31.1"]
ret3[`BayesSpace Clustering`=="k31"&`Visium Cluster`=="X4",`Visium Cluster`:="ARC31.2"]
ret3[`BayesSpace Clustering`=="k31"&`Visium Cluster`=="X11",`Visium Cluster`:="ARC31.3"]
ret3[`BayesSpace Clustering`=="k31"&`Visium Cluster`=="X22",`Visium Cluster`:="ARC31.4"]

ret3 <- melt.data.table(ret3,id.vars=c("BayesSpace Clustering","Visium Cluster","ntop"))

setnames(ret3,"variable","Mouse Subclass or Aggregate VMH/ARC Subclasses")
setnames(ret3,"value","Spearman Correlation")
setnames(ret3,"ntop","n Top Visium Cluster Markers")

ret3 <- ret3[,c(1,2,4,3,5),with=F]

fwrite(ret3,"local_supptab_assembly/sphyp-ABA23-VMHARCsubtypesCollapsed_Visium_spatreg.txt",sep='\t',quote=F)
```


### reproducibility info
```{r}
sessionInfo()
sessioninfo::session_info()
```
R version 4.4.1 RC (2024-06-06 r86719)
Platform: aarch64-apple-darwin20
Running under: macOS Sonoma 14.5

Matrix products: default
BLAS:   /System/Library/Frameworks/Accelerate.framework/Versions/A/Frameworks/vecLib.framework/Versions/A/libBLAS.dylib 
LAPACK: /Library/Frameworks/R.framework/Versions/4.4-arm64/Resources/lib/libRlapack.dylib;  LAPACK version 3.12.0

locale:
[1] en_US.UTF-8/en_US.UTF-8/en_US.UTF-8/C/en_US.UTF-8/en_US.UTF-8

time zone: America/Chicago
tzcode source: internal

attached base packages:
[1] stats4    stats     graphics  grDevices utils     datasets  methods   base     

other attached packages:
 [1] Biostrings_2.72.1   GenomeInfoDb_1.40.1 XVector_0.44.0      IRanges_2.38.1     
 [5] S4Vectors_0.42.1    BiocGenerics_0.50.0 default_1.0.0       colorout_1.3-0.2   
 [9] data.table_1.15.4   rlang_1.1.4        

loaded via a namespace (and not attached):
 [1] crayon_1.5.3            httr_1.4.7              cli_3.6.3              
 [4] knitr_1.48              xfun_0.45               UCSC.utils_1.0.0       
 [7] jsonlite_1.8.8          rprojroot_2.0.4         htmltools_0.5.8.1      
[10] rmarkdown_2.27          evaluate_0.24.0         fastmap_1.2.0          
[13] yaml_2.3.9              compiler_4.4.1          here_1.0.1             
[16] rstudioapi_0.16.0       digest_0.6.36           R6_2.5.1               
[19] GenomeInfoDbData_1.2.12 tools_4.4.1             zlibbioc_1.50.0        
> sessioninfo::session_info()
─ Session info ───────────────────────────────────────────────────────────────────
 setting  value
 version  R version 4.4.1 RC (2024-06-06 r86719)
 os       macOS Sonoma 14.5
 system   aarch64, darwin20
 ui       RStudio
 language (EN)
 collate  en_US.UTF-8
 ctype    en_US.UTF-8
 tz       America/Chicago
 date     2024-08-04
 rstudio  2024.07.0-daily+219 Cranberry Hibiscus (desktop)
 pandoc   3.1.11 @ /Applications/RStudio.app/Contents/Resources/app/quarto/bin/tools/aarch64/ (via rmarkdown)

─ Packages ───────────────────────────────────────────────────────────────────────
 ! package          * version date (UTC) lib source
   BiocGenerics     * 0.50.0  2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   Biostrings       * 2.72.1  2024-06-02 [1] Bioconductor 3.19 (R 4.4.0)
 P cli                3.6.3   2024-06-21 [2] CRAN (R 4.4.0)
   colorout         * 1.3-0.2 2024-05-01 [1] Github (jalvesaq/colorout@c6113a2)
   crayon             1.5.3   2024-06-20 [1] CRAN (R 4.4.0)
   data.table       * 1.15.4  2024-03-30 [2] CRAN (R 4.4.0)
   default          * 1.0.0   2017-08-07 [1] CRAN (R 4.4.0)
   digest             0.6.36  2024-06-23 [1] CRAN (R 4.4.0)
   evaluate           0.24.0  2024-06-10 [1] CRAN (R 4.4.0)
   fastmap            1.2.0   2024-05-15 [1] CRAN (R 4.4.0)
   GenomeInfoDb     * 1.40.1  2024-05-24 [1] Bioconductor 3.19 (R 4.4.0)
   GenomeInfoDbData   1.2.12  2024-05-01 [1] Bioconductor
   here               1.0.1   2020-12-13 [1] CRAN (R 4.4.0)
   htmltools          0.5.8.1 2024-04-04 [1] CRAN (R 4.4.0)
   httr               1.4.7   2023-08-15 [1] CRAN (R 4.4.0)
   IRanges          * 2.38.1  2024-07-03 [1] Bioconductor 3.19 (R 4.4.1)
   jsonlite           1.8.8   2023-12-04 [1] CRAN (R 4.4.0)
   knitr              1.48    2024-07-07 [1] CRAN (R 4.4.1)
   R6                 2.5.1   2021-08-19 [1] CRAN (R 4.4.0)
 P rlang            * 1.1.4   2024-06-04 [2] CRAN (R 4.4.1)
   rmarkdown          2.27    2024-05-17 [1] CRAN (R 4.4.0)
   rprojroot          2.0.4   2023-11-05 [1] CRAN (R 4.4.0)
   rstudioapi         0.16.0  2024-03-24 [1] CRAN (R 4.4.0)
   S4Vectors        * 0.42.1  2024-07-03 [1] Bioconductor 3.19 (R 4.4.1)
   sessioninfo        1.2.2   2021-12-06 [1] CRAN (R 4.4.0)
   UCSC.utils         1.0.0   2024-05-06 [1] Bioconductor 3.19 (R 4.4.0)
   xfun               0.45    2024-06-16 [1] CRAN (R 4.4.0)
   XVector          * 0.44.0  2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)
   yaml               2.3.9   2024-07-05 [1] CRAN (R 4.4.0)
   zlibbioc           1.50.0  2024-04-30 [1] Bioconductor 3.19 (R 4.4.0)

 [1] /Library/Frameworks/R.framework/Versions/4.4-arm64/Resources/library
 [2] /Users/bmulvey/Library/R/arm64/4.4/library

 P ── Loaded and on-disk path mismatch.

──────────────────────────────────────────────────────────────────────────────────
